authorization

+++ app/Http/Controllers/IdeaController.php
@@ -14,6 +14,8 @@ class IdeaController extends Controller
             'content' => 'required|min:3|max:240'
         ]);
 
+        $validated['user_id'] = auth()->id();
+
         Idea::create($validated);
 
         return redirect()->route('dashboard')->with('success', 'Idea was created successfully!');
@@ -27,6 +29,9 @@ class IdeaController extends Controller
     }
 
     public function edit(Idea $idea){
+        if(auth()->id() !== $idea->user_id){
+            abort(404, "nah uh");
+        }
 
         $editing = true;
         return view('ideas.show', compact('idea', 'editing'));
@@ -34,6 +39,9 @@ class IdeaController extends Controller
     }
 
     public function update(Idea $idea){
+        if(auth()->id() !== $idea->user_id){
+            abort(404, "nah uh");
+        }
 
         $validated = request()->validate([
             'content' => 'required|min:3|max:240'
@@ -46,6 +54,9 @@ class IdeaController extends Controller
     }
 
     public function destroy(Idea $idea){
+        if(auth()->id() !== $idea->user_id){
+            abort(404, "nah uh");
+        }
         
         $idea->delete();
